# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for personal blog.",
        "title": "Golang Restful API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/author": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "author"
                ],
                "summary": "Get list of authors.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/example.GetAuthors"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "\u003c/p\u003e[name] : Author name (max:100) \u003c/p\u003e[born] : The born year of the author (min:1000, max:2500)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "author"
                ],
                "summary": "Add new author",
                "parameters": [
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/example.PostAuthor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.OK"
                        }
                    }
                }
            }
        },
        "/author/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "author"
                ],
                "summary": "Get author info by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/example.GetAuthor"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/example.InvalidID"
                        }
                    }
                }
            },
            "put": {
                "description": "\u003c/p\u003e[name] : Author name (max:100) \u003c/p\u003e[born] : The born year of the author (min:1000, max:2500)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "author"
                ],
                "summary": "Modify author info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/example.PostAuthor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.OK"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/example.AuthorNotExist"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "author"
                ],
                "summary": "Delete author info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.OK"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/example.AuthorNotExist"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "\u003c/p\u003e[account] : (min: 6, max:12)\u003c/p\u003e[password] : (min: 6, max:12)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login and get new JWT token",
                "parameters": [
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/example.PostLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.OK"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/example.ErrorPasswdWrong"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "\u003c/p\u003e[account] : (min: 6, max:12) \u003c/p\u003e[name] : (max:12) \u003c/p\u003e[password] : (min: 6, max:12)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/example.PostRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.OK"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/example.ErrorExistUser"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "example.AuthorInfo": {
            "type": "object",
            "properties": {
                "born": {
                    "type": "integer",
                    "example": 1564
                },
                "created_on": {
                    "type": "integer",
                    "example": 1574952062
                },
                "deleted_on": {
                    "type": "integer",
                    "example": 0
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "modified_on": {
                    "type": "integer",
                    "example": 1575189888
                },
                "name": {
                    "type": "string",
                    "example": "William Shakespeare"
                }
            }
        },
        "example.AuthorNotExist": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 10002
                },
                "error_message": {
                    "type": "string",
                    "example": "Author not exist"
                },
                "result": {
                    "type": "string"
                }
            }
        },
        "example.ErrorExistUser": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 10005
                },
                "error_message": {
                    "type": "string",
                    "example": "User account already exist"
                },
                "result": {
                    "type": "string"
                }
            }
        },
...
        "example.ErrorPasswdWrong": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 10006
                },
                "error_message": {
                    "type": "string",
                    "example": "Wrong password"
                },
                "result": {
                    "type": "string"
                }
            }
        },
...
        "example.GetAuthor": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 200
                },
                "error_message": {
                    "type": "string",
                    "example": "OK"
                },
                "result": {
                    "type": "object",
                    "$ref": "#/definitions/example.AuthorInfo"
                }
            }
        },
        "example.GetAuthors": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 200
                },
                "error_message": {
                    "type": "string",
                    "example": "OK"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/example.AuthorInfo"
                    }
                }
            }
        },
        "example.InvalidID": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 10001
                },
                "error_message": {
                    "type": "string",
                    "example": "Invalid ID"
                },
                "result": {
                    "type": "string"
                }
            }
        },
        "example.OK": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer",
                    "example": 200
                },
                "error_message": {
                    "type": "string",
                    "example": "OK"
                },
                "result": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
...
        "example.PostAuthor": {
            "type": "object",
            "properties": {
                "born": {
                    "type": "integer",
                    "example": 1847
                },
                "name": {
                    "type": "string",
                    "example": "Thomas Alva Edison"
                }
            }
        },
        "example.PostLogin": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string",
                    "example": "edison1234"
                },
                "password": {
                    "type": "string",
                    "example": "test1234"
                }
            }
        },
        "example.PostRegister": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string",
                    "example": "edison1234"
                },
                "name": {
                    "type": "string",
                    "example": "edison"
                },
                "password": {
                    "type": "string",
                    "example": "test1234"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}